requirements.txt
tensorflow-gpu==2.0
pillow==6.1.0
opencv-contrib-python==4.1.1.26
tqdm==4.36.1